requirements.txt
requests
pandas
python-dotenv
psycopg2
streamlit
plotly-express
sqlalchemy>=1.4.16